server/requirements.txt
# Production dependencies with specific versions
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.8
chromadb==0.4.18
httpx==0.25.2
edge-tts==6.1.9
pydantic==2.5.0